update type signatures

+++ console_apps/calculator.py
@@ -1,14 +1,14 @@
-def add(a: float | int, b: float | int) -> float | int:
+def add(a: int | float, b: int | float) -> int | float:
     return a + b
 
 
-def subtract(a: float | int, b: float | int) -> float | int:
+def subtract(a: int | float, b: int | float) -> int | float:
     return a - b
 
 
-def multiply(a: float | int, b: float | int) -> float | int:
+def multiply(a: int | float, b: int | float) -> int | float:
     return a * b
 
 
-def divide(a: float | int, b: float | int) -> float | int:
+def divide(a: int | float, b: int | float) -> int | float:
     return a / b
